% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinEncode.R
\name{Binary_encoding}
\alias{Binary_encoding}
\title{Binary Encoding}
\usage{
Binary_encoding(data1)
}
\arguments{
\item{dataframe}{dataframe with factor/character columns}
}
\value{
dataframe with binary columns
}
\description{
Binary Encoding
}
